services/inventory_service/requirements-test.txt
pytest
pytest-django
pytest-mock
factory-boy